chore: refactoring and use mock in cleanup tests
