+++ modules/engines/engine-xsjs/src/main/java/com/sap/xsk/synchronizer/XSJSLibSynchronizerPublisherHandler.java
@@ -11,9 +11,23 @@
  */
 package com.sap.xsk.synchronizer;
 
+import com.sap.xsk.synchronizer.XSJSLibSynchronizerPathTypeResolver.ResolvedPathType;
+import com.sap.xsk.synchronizer.cleaners.XSJSLibSynchronizerCleaner;
+import com.sap.xsk.synchronizer.cleaners.XSJSLibSynchronizerDBCleaner;
+import com.sap.xsk.synchronizer.cleaners.XSJSLibSynchronizerFileCleaner;
+import org.eclipse.dirigible.commons.config.StaticObjects;
 import org.eclipse.dirigible.core.publisher.api.handlers.MetadataPublisherHandler;
+import org.eclipse.dirigible.repository.api.IRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import javax.sql.DataSource;
 
 public class XSJSLibSynchronizerPublisherHandler extends MetadataPublisherHandler {
+  private static final DataSource dataSource = (DataSource) StaticObjects.get(StaticObjects.SYSTEM_DATASOURCE);
+
+  private static final IRepository repository = (IRepository) StaticObjects.get(StaticObjects.REPOSITORY);
+
+  private static final Logger logger = LoggerFactory.getLogger(XSJSLibSynchronizerPublisherHandler.class);
 
   @Override
   public void afterPublish(String workspaceLocation, String registryLocation) {
@@ -22,7 +36,24 @@ public class XSJSLibSynchronizerPublisherHandler extends MetadataPublisherHandle
 
   @Override
   public void beforeUnpublish(String location) {
-    XSJSLibSynchronizerArtefactsCleaner cleaner = new XSJSLibSynchronizerArtefactsCleaner();
-    cleaner.cleanup(location);
+    cleanup(location);
+  }
+
+  private void cleanup(String registryPath) {
+    XSJSLibSynchronizerCleaner dbCleaner = new XSJSLibSynchronizerDBCleaner(dataSource);
+    XSJSLibSynchronizerCleaner fileCleaner = new XSJSLibSynchronizerFileCleaner(repository);
+    XSJSLibSynchronizerPathTypeResolver resolver = new XSJSLibSynchronizerPathTypeResolver();
+    ResolvedPathType type = resolver.resolveWithCollectionFirst(registryPath);
+
+    switch(type) {
+      case EXISTENT_XSJSLIB_FILE: {
+        dbCleaner.cleanup(registryPath);
+        fileCleaner.cleanup(registryPath);
+      } break;
+
+      case EXISTENT_FOLDER: dbCleaner.cleanup(registryPath); break;
+
+      default: logger.info("XSJSLibSynchronizer: Nothing to cleanup."); break;
+    }
   }
 }
