+++ modules/engines/engine-xsjs/src/test/java/com/sap/xsk/synchronizer/XSJSLibSynchronizerJobTest.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright (c) 2022 SAP SE or an SAP affiliate company and XSK contributors
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Apache License, v2.0
+ * which accompanies this distribution, and is available at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and XSK contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package com.sap.xsk.synchronizer;
+
+import com.sap.xsk.XSJSTest;
+import org.eclipse.dirigible.core.scheduler.api.ISchedulerCoreService;
+import org.eclipse.dirigible.core.scheduler.service.definition.JobDefinition;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+
+public class XSJSLibSynchronizerJobTest extends XSJSTest {
+  @Test
+  public void synchronizerJobDefinitionTest() {
+    XSJSLibSynchronizerJob job = new XSJSLibSynchronizerJob();
+    JobDefinition jobDefinition = job.getJobDefinition();
+
+    assertEquals("Unexpected XSJSLib Job Definition Name",
+        "dirigible-internal-xsk-xsjslib-synchronizer-job", jobDefinition.getName());
+
+    assertEquals("Unexpected XSJSLib Job Definition Group",
+        ISchedulerCoreService.JOB_GROUP_INTERNAL, jobDefinition.getGroup());
+
+    assertEquals("Unexpected XSJSLib Job Definition Clazz",
+        XSJSLibSynchronizerJob.class.getCanonicalName(), jobDefinition.getClazz());
+
+    assertEquals("Unexpected XSJSLib Job Definition Description",
+        "XSK XSJSLib Synchronizer Job", jobDefinition.getDescription());
+
+    assertEquals("Unexpected XSJSLib Job Definition Expression",
+        "0/55 * * * * ?", jobDefinition.getExpression());
+
+    assertTrue("Unexpected XSJSLib Job Definition Singleton Flag",
+        jobDefinition.isSingleton());
+  }
+
+  @Test
+  public void synchronizerJobTest() {
+    XSJSLibSynchronizerJob job = new XSJSLibSynchronizerJob();
+
+    assertEquals("Unexpected XSJSLib Job Name",
+        "XSK XSJSLib Synchronizer Job", job.getName());
+
+    assertEquals("Unexpected XSJSLib Job Synchronizer",
+        XSJSLibSynchronizer.class, job.getSynchronizer().getClass());
+  }
+
+}
