+++ modules/engines/engine-xsjs/src/test/java/com/sap/xsk/synchronizer/XSJSLibSynchronizerPathTypeResolverTest.java
@@ -57,20 +57,20 @@ public class XSJSLibSynchronizerPathTypeResolverTest extends XSJSTest {
   public void resolveWithCollectionFirstTest() {
     XSJSLibSynchronizerPathTypeResolver resolver = new XSJSLibSynchronizerPathTypeResolver();
     IRepository repository = (IRepository) StaticObjects.get(StaticObjects.REPOSITORY);
-    repository.createCollection("/resolveWithResourceFirstTest");
-    repository.createResource("/resolveWithResourceFirstTest/test.xsjslib");
-    repository.createResource("/resolveWithResourceFirstTest/test.html");
+    repository.createCollection("/resolveWithCollectionFirstTest");
+    repository.createResource("/resolveWithCollectionFirstTest/test.xsjslib");
+    repository.createResource("/resolveWithCollectionFirstTest/test.html");
 
     // Existent files/folders
-    ResolvedPathType type1 = resolver.resolveWithCollectionFirst("/resolveWithResourceFirstTest/test.xsjslib");
-    ResolvedPathType type2 = resolver.resolveWithCollectionFirst("/resolveWithResourceFirstTest/test.html");
-    ResolvedPathType type3 = resolver.resolveWithCollectionFirst("/resolveWithResourceFirstTest/");
-    ResolvedPathType type4 = resolver.resolveWithCollectionFirst("/resolveWithResourceFirstTest");
+    ResolvedPathType type1 = resolver.resolveWithCollectionFirst("/resolveWithCollectionFirstTest/test.xsjslib");
+    ResolvedPathType type2 = resolver.resolveWithCollectionFirst("/resolveWithCollectionFirstTest/test.html");
+    ResolvedPathType type3 = resolver.resolveWithCollectionFirst("/resolveWithCollectionFirstTest/");
+    ResolvedPathType type4 = resolver.resolveWithCollectionFirst("/resolveWithCollectionFirstTest");
 
     // Non-Existent File/folder
-    ResolvedPathType type5 = resolver.resolveWithCollectionFirst("/resolveWithResourceFirstTestNonExistent/");
-    ResolvedPathType type6 = resolver.resolveWithCollectionFirst("/resolveWithResourceFirstTestNonExistent");
-    ResolvedPathType type7 = resolver.resolveWithCollectionFirst("/resolveWithResourceFirstTest/test2.html");
+    ResolvedPathType type5 = resolver.resolveWithCollectionFirst("/resolveWithCollectionFirstTestNonExistent/");
+    ResolvedPathType type6 = resolver.resolveWithCollectionFirst("/resolveWithCollectionFirstTestNonExistent");
+    ResolvedPathType type7 = resolver.resolveWithCollectionFirst("/resolveWithCollectionFirstTest/test2.html");
 
     assertEquals("Unexpected ResolvedPathType.", ResolvedPathType.EXISTENT_XSJSLIB_FILE, type1);
     assertEquals("Unexpected ResolvedPathType.", ResolvedPathType.EXISTENT_OTHER_FILE, type2);
