+++ modules/engines/engine-xsjs/src/test/java/com/sap/xsk/synchronizer/exceptions/XSJSLibArtefactCleanerSQLExceptionTest.java
@@ -9,24 +9,18 @@
  * SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and XSK contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.sap.xsk.synchronizer;
+package com.sap.xsk.synchronizer.exceptions;
 
-import com.sap.xsk.exceptions.XSJSLibExportsGenerationSourceNotFoundException;
+import com.sap.xsk.exceptions.XSJSLibArtefactCleanerSQLException;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 
-public class XSJSLibExportsGenerationSourceNotFoundExceptionTest {
-
+public class XSJSLibArtefactCleanerSQLExceptionTest {
   @Test
-  public void exportGenerationSourceNotFoundExceptionMessageTest() {
-    XSJSLibExportsGenerationSourceNotFoundException exception = new XSJSLibExportsGenerationSourceNotFoundException("test");
+  public void artefactCleanerSQLExceptionTest() {
+    XSJSLibArtefactCleanerSQLException exception = new XSJSLibArtefactCleanerSQLException("test", new RuntimeException("test2"));
     assertEquals("Unexpected exception message", "test", exception.getMessage());
-  }
-
-  @Test
-  public void exportGenerationSourceNotFoundExceptionCauseTest() {
-    XSJSLibExportsGenerationSourceNotFoundException exception = new XSJSLibExportsGenerationSourceNotFoundException(new RuntimeException("test"));
     assertEquals("Unexpected exception cause", RuntimeException.class, exception.getCause().getClass());
   }
 }
